requirements.txt
asttokens==2.0.8
backcall==0.2.0
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
fonttools==4.37.3
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.3.1
PyQt6-Qt6==6.3.2
PyQt6-sip==13.4.0
PySide2==5.15.2.1
PySide6==6.3.2
PySide6-Addons==6.3.2
PySide6-Essentials==6.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pywin32==304
pyzmq==24.0.1
rich==12.6.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
shiboken2==5.15.2.1
shiboken6==6.3.2
six==1.16.0
sklearn==0.0
stack-data==0.5.1
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
